use pterm on output and pipe distribution logs
All logging occurs through logrus. When outputting to the console, pterm
is used. When a `task` field is present, a spinner for the task is
created. All new logs with the same task field update the spinner text.
The spinner ends when there is a log with an error field. If the field
is nil, the task has succeeded. When the error field is not nil, the task
is considered as failed.

TODO: find a pretty way to print fields.