✨ Make configuration idempotent
Now kaweezle performs the configuration every time it starts the
cluster. It means that the configuration is indempotent and will
be only changed if needed.

The configure command has also been improved by adding the
possibility to force domain name mapping configuration and to
remove the route that is now made persistent.